==> Notebooks/01_data_loading.ipynb <==
import pandas as pd
import os

def load_and_merge_csvs(folder_path):
    files = [
        os.path.join(folder_path, f)
        for f in os.listdir(folder_path)
        if f.endswith(".csv")
    ]

    df_list = [pd.read_csv(f) for f in files]
    merged_df = pd.concat(df_list, ignore_index=True)

    return merged_df



# Loading Enrolment data :-

enrolment_df = load_and_merge_csvs("../data/raw/enrolment/")

enrolment_df.head()

Checking Size of data :

enrolment_df.shape

Checking names of columns

enrolment_df.columns

# Loading Biometric Data :-

biometric_df = load_and_merge_csvs("../data/raw/biometric/")

biometric_df.head()

biometric_df.shape


biometric_df.columns


# Loading Demographic Data:-

demographic_df = load_and_merge_csvs("../data/raw/demographic/")


demographic_df.head()


demographic_df.shape

# Merging the respective CSVs

biometric_df.to_csv(
    "../data/processed/interim/biometric_raw_merged.csv",
    index=False
)

enrolment_df.to_csv(
    "../data/processed/interim/enrolment_raw_merged.csv",
    index=False
)

demographic_df.to_csv(
    "../data/processed/interim/demographic_raw_merged.csv",
    index=False
)
