==> Notebooks/03_biometric_cleaning.ipynb <==
import pandas as pd

biometric_df = pd.read_csv("../data/processed/interim/biometric_raw_merged.csv")

biometric_df.head()

biometric_df.columns


biometric_df = biometric_df.rename(columns={
    "bio_age_17_": "bio_age_17_plus"
})


biometric_df.shape

biometric_df["state"] = biometric_df["state"].str.strip().str.title()

biometric_df["district"] = biometric_df["district"].str.strip().str.title()

biometric_df["date"] = pd.to_datetime(biometric_df["date"], dayfirst=True)

biometric_df.dtypes

biometric_df = biometric_df.drop_duplicates()


biometric_df.shape

biometric_df[
    biometric_df["district"].isin(["100000", 100000]) |
    biometric_df["state"].isin(["100000", 100000])
].shape[0]


biometric_df.shape

biometric_df.isnull().sum()


biometric_df["state"].value_counts()


biometric_state_mapping = {
    "Orissa": "Odisha",
    "Pondicherry": "Puducherry",

    "West Bangal": "West Bengal",
    "Westbengal": "West Bengal",
    "West  Bengal": "West Bengal",

    "Jammu & Kashmir": "Jammu And Kashmir",

    "Chhatisgarh": "Chhattisgarh",
    "Uttaranchal": "Uttarakhand",

    "Tamilnadu": "Tamil Nadu",

    "Dadra And Nagar Haveli": "Dadra And Nagar Haveli And Daman And Diu",
    "Dadra & Nagar Haveli": "Dadra And Nagar Haveli And Daman And Diu",
    "Daman And Diu": "Dadra And Nagar Haveli And Daman And Diu",
    "Daman & Diu": "Dadra And Nagar Haveli And Daman And Diu",

    "Andaman & Nicobar Islands": "Andaman And Nicobar Islands"
}


biometric_df["state"] = biometric_df["state"].replace(biometric_state_mapping)


# -----------------------------
# Fix Hyderabad → Telangana
# -----------------------------

mask = (
    biometric_df["district"].astype(str).str.strip().str.title() == "Hyderabad"
)

affected_rows = mask.sum()

biometric_df.loc[mask, "state"] = "Telangana"

print(f"✔ Hyderabad correction applied to {affected_rows} rows")


biometric_df["state"].value_counts()


before = len(biometric_df)

biometric_df = biometric_df[
    (biometric_df["state"] != "<unset>") &
    (biometric_df["district"] != "<unset>") &
    (biometric_df["pincode"] != "<unset>")
]

after = len(biometric_df)

print(f"Removed {before - after} rows with <unset> in key columns")


biometric_df.to_csv(
    "../data/processed/cleaned/biometric_clean.csv",
    index=False
)
